Fix transaction total float64 (#65)

+++ internal/transaction/service/transaction_service.go
@@ -63,7 +63,8 @@ func (svc *transactionService) Create(p *request.CheckoutRequest) <-chan util.Re
 			return
 		}
 
-		var total = 0
+		var total = 0.0
+
 		var details []transaction.ProductDetail
 
 		for _, detail := range p.ProductDetails {
@@ -110,17 +111,19 @@ func (svc *transactionService) Create(p *request.CheckoutRequest) <-chan util.Re
 				Quantity:  detail.Quantity,
 			}
 			details = append(details, d)
-			total += int(productExists.Price * float64(detail.Quantity))
+
+			total += float64(productExists.Price * float64(detail.Quantity))
+
 		}
 
-		if float64(total) > p.Paid {
+		if total > p.Paid {
 			result <- util.ResultErr{
 				Error: errs.PaidIsNotEnough,
 			}
 			return
 		}
 
-		validChange := p.Paid - float64(total)
+		validChange := p.Paid - total
 		if validChange != *p.Change {
 			result <- util.ResultErr{
 				Error: errs.ChangeIsNotRight,
